+++ src/monsters/pokemon_creation.c
@@ -34,7 +34,7 @@ void pokemon_init(Pokemon * new_pok, int level, int level_min, int level_max) {
     new_pok->num_hidden_conditions = 0;
     
     new_pok->hidden_conditions = malloc(sizeof(Condition));
-    new_pok->hidden_condition_values = malloc(sizeof(int8_t));
+    new_pok->hidden_condition_values = malloc(sizeof(int64_t));
 
     calculate_stats(new_pok, level, level_min, level_max);
     new_pok->currentHP = new_pok->maxHP;
